Add possibility of using multiple losses
Allow multiple losses in the supervised trainer. Losses are provided in
a list, and an additional argument is passed to the trainer that
specifies how the losses should be scaled.
Currently, this functionality is only added in the supervised trainer,
passing different losses via the commandline to the train_model.py
script is not supported.

Furthermore, introduce a function `get_batch_data` in the supervised
trainer that returns the data for the current batch, instead of
generating it in the `_train_epoches` function itself. This allows this
function to be used by classes that inherit from the SupervisedTrainer
but use different types of targets.