Fix improper probability distribution when pondering
For every ponder step we would pass the hidden state through an output layer and apply log_softmax, after we would take a weighted average of these log_softmaxes.
Now, we move this layer outside of the pondering mechanism.
So we take the average over rnn states, which gets forwarded through this layer and activation only once.